Renamed functions again.
