+++ VACARM.Backend.NET8/VACARM.Infrastructure/Functions/DeviceFunctions.cs
@@ -1,86 +1,120 @@
 ﻿using AudioSwitcher.AudioApi;
+using NAudio.CoreAudioApi;
+using VACARM.Domain.Models;
 
 namespace VACARM.Infrastructure.Functions
 {
-  internal static class DeviceFunctions<TDevice>
-    where TDevice :
-    Device
+  internal static class DeviceFunctions<TDeviceModel> 
+    where TDeviceModel :
+    DeviceModel
   {
     #region Parameters
 
-    private const DeviceState PresentDeviceState =
-      DeviceState.Active
-      | DeviceState.Disabled
-      | DeviceState.Unplugged;
+    internal readonly static Func<TDeviceModel, bool> IsAbsent =
+      (TDeviceModel x) => !x.IsPresent;
 
-    internal readonly static Func<TDevice, bool> IsAbsent =
-      (TDevice x) => x.State != PresentDeviceState;
+    internal readonly static Func<TDeviceModel, bool> IsCapture =
+      (TDeviceModel x) => x.IsCapture;
 
-    internal readonly static Func<TDevice, bool> IsCapture =
-      (TDevice x) => x.IsCaptureDevice;
+    internal readonly static Func<TDeviceModel, bool> IsCommunications =
+      (TDeviceModel x) => x.Role == "Communications";
 
-    internal readonly static Func<TDevice, bool> IsDefault =
-      (TDevice x) => x.IsDefaultDevice;
+    internal readonly static Func<TDeviceModel, bool> IsConsole =
+      (TDeviceModel x) => x.Role == "Console";
 
-    internal readonly static Func<TDevice, bool> IsDefaultCommunications =
-      (TDevice x) => x.IsDefaultCommunicationsDevice;
+    internal readonly static Func<TDeviceModel, bool> IsDefault =
+      (TDeviceModel x) => x.IsDefault;
 
-    internal readonly static Func<TDevice, bool> IsDisabled =
-      (TDevice x) => x.State == DeviceState.Disabled;
+    internal readonly static Func<TDeviceModel, bool> IsDisabled =
+      (TDeviceModel x) => !x.IsEnabled;
 
-    internal readonly static Func<TDevice, bool> IsDuplex =
-      (TDevice x) => x.IsCaptureDevice == x.IsPlaybackDevice;
+    internal readonly static Func<TDeviceModel, bool> IsDuplex =
+      (TDeviceModel x) => x.IsDuplex;
 
-    internal readonly static Func<TDevice, bool> IsEnabled =
-      (TDevice x) => x.State != DeviceState.Disabled;
+    internal readonly static Func<TDeviceModel, bool> IsEnabled =
+      (TDeviceModel x) => x.IsEnabled;
 
-    internal readonly static Func<TDevice, bool> IsMuted =
-      (TDevice x) => x.IsMuted;
+    internal readonly static Func<TDeviceModel, bool> IsMultimedia =
+      (TDeviceModel x) => x.Role == "Multimedia";
 
-    internal readonly static Func<TDevice, bool> IsPresent =
-      (TDevice x) => x.State == PresentDeviceState;
+    internal readonly static Func<TDeviceModel, bool> IsMuted =
+      (TDeviceModel x) => x.IsMuted;
 
-    internal readonly static Func<TDevice, bool> IsPlayback =
-      (TDevice x) => x.IsPlaybackDevice;
+    internal readonly static Func<TDeviceModel, bool> IsPresent =
+      (TDeviceModel x) => x.IsPresent;
 
-    internal readonly static Func<TDevice, bool> IsUnmuted =
-      (TDevice x) => !x.IsMuted;
+    internal readonly static Func<TDeviceModel, bool> IsRender =
+      (TDeviceModel x) => x.IsRender;
+
+    internal readonly static Func<TDeviceModel, bool> IsUnmuted =
+      (TDeviceModel x) => !x.IsMuted;
 
     #endregion
 
     #region Logic
 
-    private static Guid ToGuid(string id)
-    {
-      return new Guid(id);
-    }
-
     /// <summary>
-    /// Match a <typeparamref name="TDevice"/> ID.
+    /// Get a <typeparamref name="DeviceModel"/>.
     /// </summary>
     /// <param name="id">The ID</param>
-    /// <returns>The function</returns>
-    internal static Func<TDevice, bool> ContainsId(string id)
+    /// <param name="mMDevice">The device</param>
+    /// <param name="device">The device</param>
+    /// <param name="role">The role</param>
+    /// <returns>The device model.</returns>
+    internal static TDeviceModel GetDeviceModel
+    (
+      uint id,
+      MMDevice mMDevice,
+      Device? device,
+      string? role
+    )
     {
-      return (TDevice item) => item.Id == ToGuid(id);
+      return new DeviceModel
+        (
+          id,
+          mMDevice.ID,
+          mMDevice.DeviceFriendlyName,
+          MMDeviceFunctions<MMDevice>.IsCapture(mMDevice),
+          (bool?)CoreAudioDeviceFunctions<Device>.IsDefault(device),
+          MMDeviceFunctions<MMDevice>.IsDisabled(mMDevice),
+          (bool?)CoreAudioDeviceFunctions<Device>.IsMuted(device),
+          MMDeviceFunctions<MMDevice>.IsPresent(mMDevice),
+          MMDeviceFunctions<MMDevice>.IsRender(mMDevice),
+          role
+        ) as TDeviceModel;
     }
 
     /// <summary>
-    /// Match an enumerable of <typeparamref name="TDevice"/> ID(s).
+    /// Match a <typeparamref name="TDeviceModel"/> actual ID.
     /// </summary>
-    /// <param name="startId">The first ID</param>
-    /// <param name="endId">The last ID</param>
+    /// <param name="actualId">The actual ID</param>
     /// <returns>The function</returns>
-    internal static Func<TDevice, bool> ContainsIdEnumerable
-    (IEnumerable<string> idEnumerable)
+    internal static Func<TDeviceModel, bool> ContainsActualId(string actualId)
     {
-      return (TDevice item) =>
-      {
-        string id = item.Id
-          .ToString();
+      return (TDeviceModel x) => x.ActualId == actualId;
+    }
 
-        return idEnumerable.Contains(id);
-      };
+    /// <summary>
+    /// Match a <typeparamref name="TDeviceModel"/> name.
+    /// </summary>
+    /// <param name="name">The name</param>
+    /// <returns>The function</returns>
+    internal static Func<TDeviceModel, bool> ContainsName(string name)
+    {
+      return (TDeviceModel x) => x.Name
+        .ToLower()
+        .Contains(name.ToLower());
+    }
+
+    /// <summary>
+    /// Match a <typeparamref name="TDeviceModel"/> role.
+    /// </summary>
+    /// <param name="role">The role</param>
+    /// <returns>The function</returns>
+    internal static Func<TDeviceModel, bool> ContainsRole(string role)
+    {
+      return (TDeviceModel x) => x.Role
+        .ToLower() == role.ToLower();
     }
 
     #endregion
